+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/crypto/tasks/UploadKeysTask.kt
@@ -23,7 +23,6 @@ import im.vector.matrix.android.internal.crypto.model.rest.KeysUploadResponse
 import im.vector.matrix.android.internal.crypto.model.rest.RestDeviceInfo
 import im.vector.matrix.android.internal.network.executeRequest
 import im.vector.matrix.android.internal.task.Task
-import im.vector.matrix.android.internal.util.convertToUTF8
 import org.greenrobot.eventbus.EventBus
 import timber.log.Timber
 import javax.inject.Inject
@@ -33,9 +32,8 @@ internal interface UploadKeysTask : Task<UploadKeysTask.Params, KeysUploadRespon
             // the device keys to send.
             val deviceKeys: RestDeviceInfo?,
             // the one-time keys to send.
-            val oneTimeKeys: JsonDict?,
-            // the explicit device_id to use for upload (default is to use the same as that used during auth).
-            val deviceId: String)
+            val oneTimeKeys: JsonDict?
+    )
 }
 
 internal class DefaultUploadKeysTask @Inject constructor(
@@ -44,8 +42,6 @@ internal class DefaultUploadKeysTask @Inject constructor(
 ) : UploadKeysTask {
 
     override suspend fun execute(params: UploadKeysTask.Params): KeysUploadResponse {
-        val encodedDeviceId = convertToUTF8(params.deviceId)
-
         val body = KeysUploadBody(
                 deviceKeys = params.deviceKeys,
                 oneTimeKeys = params.oneTimeKeys
@@ -54,11 +50,7 @@ internal class DefaultUploadKeysTask @Inject constructor(
         Timber.i("## Uploading device keys -> $body")
 
         return executeRequest(eventBus) {
-            apiCall = if (encodedDeviceId.isBlank()) {
-                cryptoApi.uploadKeys(body)
-            } else {
-                cryptoApi.uploadKeys(encodedDeviceId, body)
-            }
+            apiCall = cryptoApi.uploadKeys(body)
         }
     }
 }